Proyecto Final/requirements.txt
flask
flask-sqlalchemy
pycryptodome
requests